+++ data/storeDEPRECATED/mongo/mongo_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/globalsign/mgo/bson"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	. "github.com/onsi/gomega/gstruct"
 
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/storeDEPRECATED"
@@ -155,11 +156,38 @@ var _ = Describe("Mongo", func() {
 	})
 
 	Context("with a new store", func() {
+		var mgoSession *mgo.Session
+		var mgoCollection *mgo.Collection
+
 		BeforeEach(func() {
 			var err error
 			store, err = mongo.NewStore(config, logger)
 			Expect(err).ToNot(HaveOccurred())
 			Expect(store).ToNot(BeNil())
+			mgoSession = storeStructuredMongoTest.Session().Copy()
+			mgoCollection = mgoSession.DB(config.Database).C(config.CollectionPrefix + "deviceData")
+		})
+
+		AfterEach(func() {
+			if mgoSession != nil {
+				mgoSession.Close()
+			}
+		})
+
+		Context("EnsureIndexes", func() {
+			It("returns successfully", func() {
+				Expect(store.EnsureIndexes()).To(Succeed())
+				indexes, err := mgoCollection.Indexes()
+				Expect(err).ToNot(HaveOccurred())
+				Expect(indexes).Should(ConsistOf(
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("_id")}),
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("_userId", "_active", "_schemaVersion", "-time"), "Background": Equal(true), "Name": Equal("UserIdTypeWeighted")}),
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("origin.id", "type", "-deletedTime", "_active"), "Background": Equal(true), "Name": Equal("OriginId")}),
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("type", "uploadId"), "Background": Equal(true), "Name": Equal("typeUploadId")}),
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("uploadId", "type", "-deletedTime", "_active"), "Background": Equal(true), "Name": Equal("UploadId")}),
+					MatchFields(IgnoreExtras, Fields{"Key": ConsistOf("uploadId"), "Background": Equal(true), "Unique": Equal(true), "Name": Equal("UniqueUploadId"), "PartialFilter": HaveKeyWithValue("type", "upload")}),
+				))
+			})
 		})
 
 		Context("NewDataSession", func() {
@@ -170,18 +198,9 @@ var _ = Describe("Mongo", func() {
 		})
 
 		Context("with a new session", func() {
-			var mgoSession *mgo.Session
-			var mgoCollection *mgo.Collection
-
 			BeforeEach(func() {
 				session = store.NewDataSession()
 				Expect(session).ToNot(BeNil())
-			})
-
-			AfterEach(func() {
-				if mgoSession != nil {
-					mgoSession.Close()
-				}
 			})
 
 			Context("with persisted data sets", func() {
@@ -436,6 +455,13 @@ var _ = Describe("Mongo", func() {
 						It("returns an error if the data set with the same id already exists", func() {
 							Expect(session.CreateDataSet(ctx, dataSet)).To(Succeed())
 							Expect(session.CreateDataSet(ctx, dataSet)).To(MatchError("unable to create data set; data set already exists"))
+						})
+
+						It("returns an error if the data set with the same uploadId (but different userId) already exists", func() {
+							dataSet.UserID = pointer.FromString("differentUser")
+							Expect(session.CreateDataSet(ctx, dataSet)).To(Succeed())
+							Expect(session.CreateDataSet(ctx, dataSet)).To(MatchError("unable to create data set; data set already exists"))
+							dataSet.UserID = pointer.FromString("")
 						})
 
 						It("sets the created time", func() {