+++ data/types/food/nutrition.go
@@ -5,11 +5,17 @@ import (
 	"github.com/tidepool-org/platform/structure"
 )
 
+const (
+	EstimatedAbsorptionDurationSecondsMaximum = 86400
+	EstimatedAbsorptionDurationSecondsMinimum = 0
+)
+
 type Nutrition struct {
-	Carbohydrate *Carbohydrate `json:"carbohydrate,omitempty" bson:"carbohydrate,omitempty"`
-	Energy       *Energy       `json:"energy,omitempty" bson:"energy,omitempty"`
-	Fat          *Fat          `json:"fat,omitempty" bson:"fat,omitempty"`
-	Protein      *Protein      `json:"protein,omitempty" bson:"protein,omitempty"`
+	EstimatedAbsorptionDuration *int          `json:"estimatedAbsorptionDuration,omitempty" bson:"estimatedAbsorptionDuration,omitempty"`
+	Carbohydrate                *Carbohydrate `json:"carbohydrate,omitempty" bson:"carbohydrate,omitempty"`
+	Energy                      *Energy       `json:"energy,omitempty" bson:"energy,omitempty"`
+	Fat                         *Fat          `json:"fat,omitempty" bson:"fat,omitempty"`
+	Protein                     *Protein      `json:"protein,omitempty" bson:"protein,omitempty"`
 }
 
 func ParseNutrition(parser structure.ObjectParser) *Nutrition {
@@ -26,6 +32,7 @@ func NewNutrition() *Nutrition {
 }
 
 func (n *Nutrition) Parse(parser structure.ObjectParser) {
+	n.EstimatedAbsorptionDuration = parser.Int("estimatedAbsorptionDuration")
 	n.Carbohydrate = ParseCarbohydrate(parser.WithReferenceObjectParser("carbohydrate"))
 	n.Energy = ParseEnergy(parser.WithReferenceObjectParser("energy"))
 	n.Fat = ParseFat(parser.WithReferenceObjectParser("fat"))
@@ -33,6 +40,7 @@ func (n *Nutrition) Parse(parser structure.ObjectParser) {
 }
 
 func (n *Nutrition) Validate(validator structure.Validator) {
+	validator.Int("estimatedAbsorptionDuration", n.EstimatedAbsorptionDuration).InRange(EstimatedAbsorptionDurationSecondsMinimum, EstimatedAbsorptionDurationSecondsMaximum)
 	if n.Carbohydrate != nil {
 		n.Carbohydrate.Validate(validator.WithReference("carbohydrate"))
 	}