+++ user/service/service.go
@@ -339,6 +339,13 @@ func (s *Service) initializeConfirmationStore() error {
 	}
 	s.confirmationStore = store
 
+	s.Logger().Debug("Ensuring confirmation store indexes")
+
+	err = s.confirmationStore.EnsureIndexes()
+	if err != nil {
+		return errors.Wrap(err, "unable to ensure confirmation store indexes")
+	}
+
 	return nil
 }
 
@@ -483,6 +490,13 @@ func (s *Service) initializeUserStructuredStore() error {
 		return errors.Wrap(err, "unable to create user structured store")
 	}
 	s.userStructuredStore = userStructuredStore
+
+	s.Logger().Debug("Ensuring user structured store indexes")
+
+	err = s.userStructuredStore.EnsureIndexes()
+	if err != nil {
+		return errors.Wrap(err, "unable to ensure user structured store indexes")
+	}
 
 	return nil
 }